+++ .vscode/launch.json
@@ -15,7 +15,14 @@
             "request":       "launch",
             "name":          "Launch Build",
             "program":       "${workspaceRoot}\\App.Client\\node_modules\\@surface\\compiler\\index.js",
-            //"runtimeArgs":   ["--preserve-symlinks"],
+            "cwd":           "${workspaceRoot}\\App.Client",
+            "outFiles": []
+        },
+        {
+            "type":          "node",
+            "request":       "launch",
+            "name":          "Launch Build All",
+            "program":       "${workspaceRoot}\\tasks\\build",
             "cwd":           "${workspaceRoot}\\App.Client",
             "outFiles": []
         },
@@ -23,18 +30,19 @@
             "type":     "node",
             "request":  "launch",
             "name":     "Launch Linking",
-            "program":  "F:\\Projects\\Surface-Workbench\\tasks\\link",
+            "program":  "${workspaceRoot}\\tasks\\link",
             "args":     ["r"],
             "cwd":      "${workspaceRoot}\\tasks\\",
             "outFiles": []
         },
+        
         {
             "type":     "node",
             "request":  "launch",
             "name":     "Launch Publish",
-            "program":  "F:\\Projects\\Surface-Workbench\\tasks\\publish",
+            "program":  "${workspaceRoot}\\tasks\\publish",
             "cwd":      "${workspaceRoot}\\tasks\\",
-            "outFiles": []
+            "console":  "integratedTerminal"
         },
         {
             "type": "chrome",
