+++ tasks/build.js
@@ -1,14 +1,18 @@
 const Path    = require('path');
-const execute = require('./execute');
+const Common  = require('./common');
 const modules = require('./modules');
 
-let surfaceRoot = Path.resolve(__dirname, '../Surface/source');
+let root = Path.resolve(__dirname, '../Surface/source');
 
 let commands = [];
 
-commands.push(execute(`Compiling server`, `tsc -p ${Path.resolve(__dirname, '../App.Server/start')}`));
+commands.push(Common.execute(`Compiling server`, `tsc -p ${Path.resolve(__dirname, '../App.Server/start')}`));
 
 for (let $module of modules)
-    commands.push(execute(`Compiling ${$module.name}`, `tsc -p ${Path.normalize(`${surfaceRoot}/${$module.name}`)}`));
+{
+    let source = Path.normalize(Path.join(root, $module.name));
+    Common.cleanup(source, /\.(js(\.map)?|d\.ts)$/, /@types|node_modules/);
+    commands.push(Common.execute(`Compiling ${$module.name}`, `tsc -p ${source}`));
+}
 
 Promise.all(commands).then(() => console.log('\nDone!'));
