+++ tasks/versions.json
@@ -1,1 +1,1 @@
-{"@surface/common":"0.1.0","@surface/compiler":"0.1.0","@surface/core":"0.1.0","@surface/custom-element":"0.1.0","@surface/enumerable":"0.1.0","@surface/html-template-plugin":"0.1.0"}
+{"@surface/common":"0.0.1","@surface/compiler":"0.0.1","@surface/core":"0.0.1","@surface/custom-element":"0.0.1","@surface/enumerable":"0.0.1","@surface/html-template-plugin":"0.0.1"}
